+++ hydromt/io.py
@@ -11,6 +11,7 @@ import geopandas as gpd
 import numpy as np
 import pandas as pd
 import pyproj
+import rioxarray
 import xarray as xr
 from shapely.geometry import box
 from shapely.geometry.base import GEOMETRY_TYPES
@@ -63,16 +64,14 @@ def open_raster(
     data : DataArray
         DataArray
     """
-    kwargs.update(
-        masked=mask_nodata, default_name="data", engine="rasterio", chunks=chunks
-    )
+    kwargs.update(masked=mask_nodata, default_name="data", chunks=chunks)
     if not mask_nodata:  # if mask_and_scale by default True in xarray ?
         kwargs.update(mask_and_scale=False)
     if isinstance(filename, io.IOBase):  # file-like does not handle chunks
         logger.warning("Removing chunks to read and load remote data.")
         kwargs.pop("chunks")
     # keep only 2D DataArray
-    da = xr.open_dataarray(filename, **kwargs).squeeze(drop=True)
+    da = rioxarray.open_rasterio(filename, **kwargs).squeeze(drop=True)
     # set missing _FillValue
     if mask_nodata:
         da.raster.set_nodata(np.nan)
@@ -377,12 +376,14 @@ def open_geodataset(
     logger=logger,
     **kwargs,
 ):
-    """Open point location GIS file and timeseries file combine a single xarray.Dataset.
+    """Open and combine geometry location GIS file and timeseries file in a xr.Dataset.
 
     Arguments
     ---------
     fn_locs: path, str
-        Path to point location file, see :py:meth:`geopandas.read_file` for options.
+        Path to geometry location file, see :py:meth:`geopandas.read_file` for options.
+        For point location, the file can also be a csv, parquet, xls(x) or xy file,
+        see :py:meth:`hydromt.io.open_vector_from_table` for options.
     fn_data: path, str
         Path to data file of which the index dimension which should match the geospatial
         coordinates index.
@@ -405,7 +406,7 @@ def open_geodataset(
         CRS mis-matches are resolved if given a GeoSeries or GeoDataFrame.
         Cannot be used with bbox.
     **kwargs:
-        Key-word argume
+        Key-word argument
     logger : logger object, optional
         The logger object used for logging messages. If not provided, the default
         logger will be used.
@@ -417,8 +418,11 @@ def open_geodataset(
     """
     if not isfile(fn_locs):
         raise IOError(f"GeoDataset point location file not found: {fn_locs}")
+    # For filetype [], only point geometry is supported
+    filetype = str(fn_locs).split(".")[-1].lower()
+    if filetype in ["csv", "parquet", "xls", "xlsx", "xy"]:
+        kwargs.update(assert_gtype="Point")
     # read geometry file
-    kwargs.update(assert_gtype="Point")
     gdf = open_vector(fn_locs, crs=crs, bbox=bbox, geom=geom, **kwargs)
     if index_dim is None:
         index_dim = gdf.index.name if gdf.index.name is not None else "index"